decimal to double support for logical functions support for logical constants

+++ MacroRunner/Compiler/Formulas/ExcelFormulaOperations.cs
@@ -4,20 +4,20 @@ namespace MacroRunner.Compiler.Formulas
     {
         public static int Add(int a, int b) => a + b;
         
-        public static decimal Add(decimal a, decimal b) => a + b;
+        public static double Add(double a, double b) => a + b;
 
         public static dynamic Add(dynamic a, dynamic b) => a + b;
 
         public static int Subtract(int a, int b) => a - b;
 
-        public static decimal Subtract(decimal a, decimal b) => a - b;
+        public static double Subtract(double a, double b) => a - b;
 
         public static int Multiply(int a, int b) => a * b;
 
-        public static decimal Multiply(decimal a, decimal b) => a * b;
+        public static double Multiply(double a, double b) => a * b;
 
-        public static decimal Divide(decimal a, decimal b) => a / b;
+        public static double Divide(double a, double b) => a / b;
 
-        public static decimal Modulo(decimal a, decimal b) => a % b;
+        public static double Modulo(double a, double b) => a % b;
     }
 }
